Now able to Play and Reset the Game
In this commit I wrote the function to be able to check for a 'win' ,
and once the game is won it will indicate the the over status is true and
prompt with a play again button. If you click this button it will reset
cells and create a new game.